git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@3155 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/File__Analyze.cpp
@@ -1863,7 +1863,7 @@ void File__Analyze::Finish (File__Analyze* Parser)
     if (File_Offset+Buffer_Offset+Element_Size>=File_Size)
     {
         Element_Size=0;
-        Parser->Buffer_Offset=Parser->File_Size-Parser->File_Offset;
+        Parser->Buffer_Offset=(size_t)(Parser->File_Size-Parser->File_Offset);
     }
 
     Parser->Finish();
